+++ changelog.txt
@@ -1,3 +1,8 @@
+2025/12/16 Version 1.11 (v1.11.0):
+ * Viewer: Open the gene legend in a separate window option added.
+ * Viewer: Improved gene legend scaling.
+
+
 2025/12/09 Version 1.10 Patch (v1.10.3):
  * General: Docs for image transforming with LineStuffUp package fix.
  * Extract: Giving tile positions for numpy raw files is now easier.
